model_linebot/requirements.txt
flask
tensorflow
line-bot-sdk
uwsgi
